fix(doctor): preserve safe-mode contract; gate late check completions
P1: --doctor short-circuit must outrun heavy static imports.

ESM static imports are evaluated before any module-body code, so the
previous `if (!isDoctorInvocation) loadSeroEnv()` guard could not
prevent ./platform/env, ./ipc, or any feature module from running its
top-level side effects (resolveStartupEnv reads profiles.json and may
trigger migration/repair). Split main.ts into:

- main.ts (5.5 kB bundle): tiny doctor-aware bootstrap; only imports
  the self-contained doctor CLI module; pulls Electron via dynamic
  import inside the doctor branch.
- app-main.ts (new): the heavy app entry, formerly the body of
  main.ts. Reachable only via `await import('./app-main')` after the
  doctor branch has been ruled out.
- build-electron.mjs: adds app-main.ts as a second entry point with
  splitting:true so esbuild emits app-main.mjs as a separate output;
  the dynamic import resolves at runtime instead of being inlined.

Result: --doctor never imports platform/env, ipc, workspace, container
infra, etc. A new main-bootstrap.test.ts pins this with an allow-list
of permitted static imports + a list of explicitly forbidden modules,
so future drift surfaces immediately.

P2: Stale check-done events after all-done.

The previous runner kicked off every check concurrently and raced
Promise.all against the budget. Late completions could still mutate
`all` and emit check-done events after `all-done`. Fixed by:

- adding a `finished` flag that the map callback checks before
  pushing/emitting; late stragglers now drop silently.
- reserving check ids in `completed` during synth so a real result
  arriving between synth and `finished=true` is also dropped.
- gating `emit()` itself so nothing but `all-done` can sneak through
  after the run has been closed.
- aborting `globalAbort` immediately on close so subprocess-based
  checks waiting on the signal release the event loop.

Also threads ctx.signal into runCommand so subprocess probes (df,
container --version, container system status, which docker, docker
info) honour the budget. Helpers' new `signal` option short-circuits
before spawn if the signal is already aborted.

Tests:
- runner.test.ts: late check-done is dropped + report contains only
  the synth fail; the abort signal is observably aborted post-budget.
- main-bootstrap.test.ts (new): allow-list of static imports + check
  for forbidden path fragments + assertion that the dynamic
  ./app-main import is present.

Build verified: main.mjs is 5.5 kB, app-main.mjs is 1.2 MB, dynamic
import emits a runtime `import('./app-main.mjs')`.
